New upstream version 1.8.1

+++ src/main.c
@@ -7,7 +7,7 @@
  */
 /*
  * Copyright (C) 1997 Robey Pointer
- * Copyright (C) 1999 - 2016 Eggheads Development Team
+ * Copyright (C) 1999 - 2017 Eggheads Development Team
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -69,6 +69,7 @@
 #  define UAC_NOPRINT 0x00000001        /* Don't report unaligned fixups */
 #endif
 
+#include "version.h"
 #include "chan.h"
 #include "modules.h"
 #include "tandem.h"
@@ -103,8 +104,11 @@ static char **argv;
  * modified versions of this bot.
  */
 
-char egg_version[1024] = "1.8.0";
-int egg_numver = 1080000;
+char egg_version[1024] = EGG_STRINGVER;
+int egg_numver = EGG_NUMVER;
+#ifdef EGG_PATCH
+char egg_patch[] = EGG_PATCH;
+#endif
 
 char notify_new[121] = "";      /* Person to send a note to for new users */
 int default_flags = 0;          /* Default user flags                     */
@@ -121,7 +125,6 @@ char helpdir[121] = "help/";            /* Directory of help files  */
 char textdir[121] = "text/";            /* Directory for text files */
 
 int keep_all_logs = 0;                  /* Never erase logfiles?    */
-char logfile_suffix[21] = ".%d%b%Y";    /* Format of logfile suffix */
 int switch_logfiles_at = 300;           /* When to switch logfiles  */
 
 time_t online_since;    /* time that the bot was started */
@@ -134,7 +137,6 @@ int notify_users_at = 0; /* Minutes past the hour to notify users of notes? */
 
 char version[81];    /* Version info (long form)  */
 char ver[41];        /* Version info (short form) */
-char egg_xtra[2048]; /* Patch info                */
 
 int do_restart = 0;       /* .restart has been called, restart ASAP */
 int resolve_timeout = 15; /* Hostname/address lookup timeout        */
@@ -265,13 +267,17 @@ static void write_debug()
      *       _not_ safe <cybah>
      */
     x = creat("DEBUG.DEBUG", 0644);
-    setsock(x, SOCK_NONSOCK);
     if (x >= 0) {
+      setsock(x, SOCK_NONSOCK);
       strncpyz(s, ctime(&now), sizeof s);
       dprintf(-x, "Debug (%s) written %s\n", ver, s);
       dprintf(-x, "Please report problem to bugs@eggheads.org\n");
       dprintf(-x, "after a visit to http://www.eggheads.org/bugzilla/\n");
-      dprintf(-x, "Full Patch List: %s\n", egg_xtra);
+#ifdef EGG_PATCH
+      dprintf(-x, "Patch level: %s\n", egg_patch);
+#else
+      dprintf(-x, "Patch level: %s\n", "stable");
+#endif
       dprintf(-x, "Context: ");
       cx_ptr = cx_ptr & 15;
       for (y = ((cx_ptr + 1) & 15); y != cx_ptr; y = ((y + 1) & 15))
@@ -296,7 +302,11 @@ static void write_debug()
   } else {
     strncpyz(s, ctime(&now), sizeof s);
     dprintf(-x, "Debug (%s) written %s\n", ver, s);
-    dprintf(-x, "Full Patch List: %s\n", egg_xtra);
+#ifdef EGG_PATCH
+    dprintf(-x, "Patch level: %s\n", egg_patch);
+#else
+    dprintf(-x, "Patch level: %s\n", "stable");
+#endif
 #ifdef STATIC
     dprintf(-x, "STATICALLY LINKED\n");
 #endif
@@ -310,8 +320,7 @@ static void write_debug()
     dprintf(-x, "Tcl version: %s (header version %s)\n",
             ((interp) && (Tcl_Eval(interp, "info patchlevel") == TCL_OK)) ?
             tcl_resultstring() : (Tcl_Eval(interp, "info tclversion") == TCL_OK) ?
-            tcl_resultstring() : "*unknown*", TCL_PATCH_LEVEL ? TCL_PATCH_LEVEL :
-            "*unknown*");
+            tcl_resultstring() : "*unknown*", TCL_PATCH_LEVEL);
 
     if (tcl_threaded())
       dprintf(-x, "Tcl is threaded\n");
@@ -516,7 +525,7 @@ static void do_arg(char *s)
         printf("%s\n", version);
         if (z[0])
           printf("  (patches: %s)\n", z);
-        printf("Configured with: " EGG_AC_ARGS "\n");
+        printf("Configure flags: " EGG_AC_ARGS "\n");
         printf("Compiled with: ");
 #ifdef IPV6
         printf("IPv6, ");
@@ -550,7 +559,7 @@ void backup_userfile(void)
 }
 
 /* Timer info */
-static int lastmin = 99;
+static time_t lastmin;
 static time_t then;
 static struct tm nowtm;
 
@@ -562,6 +571,7 @@ static void core_secondly()
 {
   static int cnt = 0;
   int miltime;
+  time_t nowmins;
 
   do_check_timers(&utimer);     /* Secondly timers */
   cnt++;
@@ -577,20 +587,21 @@ static void core_secondly()
     }
   }
   egg_memcpy(&nowtm, localtime(&now), sizeof(struct tm));
-  if (nowtm.tm_min != lastmin) {
+  nowmins = time(NULL) / 60;
+  if (nowmins > lastmin) {
     int i = 0;
 
     /* Once a minute */
-    lastmin = (lastmin + 1) % 60;
+    ++lastmin;
     call_hook(HOOK_MINUTELY);
     check_expired_ignores();
     autolink_cycle(NULL);       /* Attempt autolinks */
     /* In case for some reason more than 1 min has passed: */
-    while (nowtm.tm_min != lastmin) {
+    while (nowmins != lastmin) {
       /* Timer drift, dammit */
-      debug2("timer: drift (lastmin=%d, now=%d)", lastmin, nowtm.tm_min);
+      debug2("timer: drift (lastmin=%d, nowmins=%d)", lastmin, nowmins);
       i++;
-      lastmin = (lastmin + 1) % 60;
+      ++lastmin;
       call_hook(HOOK_MINUTELY);
     }
     if (i > 1)
@@ -726,17 +737,6 @@ int init_language(int);
 #ifdef TLS
 int ssl_init();
 #endif
-
-static void patch(const char *str)
-{
-  char *p = strchr(egg_version, '+');
-
-  if (!p)
-    p = &egg_version[strlen(egg_version)];
-  sprintf(p, "+%s", str);
-  egg_numver++;
-  sprintf(&egg_xtra[strlen(egg_xtra)], " %s", str);
-}
 
 static inline void garbage_collect(void)
 {
@@ -899,7 +899,7 @@ int mainloop(int toplevel)
             d = d->next;
           }
           if (ok) {
-            strcpy(name, p->name);
+            strncpyz(name, p->name, sizeof name);
             if (module_unload(name, botnetnick) == NULL) {
               f = 1;
               break;
@@ -993,20 +993,20 @@ int main(int arg_c, char **arg_v)
     Context;
 #endif
 
-/* Include patch.h header for patch("...") */
-#include "patch.h"
-
   argc = arg_c;
   argv = arg_v;
 
   /* Version info! */
+#ifdef EGG_PATCH
+  egg_snprintf(&egg_version[strlen(egg_version)], sizeof egg_version, 
+               "+%s", egg_patch);
+#endif
   egg_snprintf(ver, sizeof ver, "eggdrop v%s", egg_version);
   egg_snprintf(version, sizeof version,
-               "Eggdrop v%s (C) 1997 Robey Pointer (C) 2010-2016 Eggheads",
+               "Eggdrop v%s (C) 1997 Robey Pointer (C) 1999-2017 Eggheads",
                egg_version);
   /* Now add on the patchlevel (for Tcl) */
   sprintf(&egg_version[strlen(egg_version)], " %u", egg_numver);
-  strcat(egg_version, egg_xtra);
 
 /* For OSF/1 */
 #ifdef STOP_UAC
@@ -1049,7 +1049,7 @@ int main(int arg_c, char **arg_v)
   now = time(NULL);
   chanset = NULL;
   egg_memcpy(&nowtm, localtime(&now), sizeof(struct tm));
-  lastmin = nowtm.tm_min;
+  lastmin = now / 60;
   srandom((unsigned int) (now % (getpid() + getppid())));
   init_mem();
   init_language(1);
@@ -1107,16 +1107,20 @@ int main(int arg_c, char **arg_v)
   /* Check for pre-existing eggdrop! */
   f = fopen(pid_file, "r");
   if (f != NULL) {
-    fgets(s, 10, f);
-    xx = atoi(s);
-    i = kill(xx, SIGCHLD);      /* Meaningless kill to determine if pid
-                                 * is used */
-    if (i == 0 || errno != ESRCH) {
-      printf(EGG_RUNNING1, botnetnick);
-      printf(EGG_RUNNING2, pid_file);
-      bg_send_quit(BG_ABORT);
-      exit(1);
+    if (fgets(s, 10, f) != NULL) {
+      xx = atoi(s);
+      i = kill(xx, SIGCHLD);      /* Meaningless kill to determine if pid
+                                   * is used */
+      if (i == 0 || errno != ESRCH) {
+        printf(EGG_RUNNING1, botnetnick);
+        printf(EGG_RUNNING2, pid_file);
+        bg_send_quit(BG_ABORT);
+        exit(1);
+      }
+    } else {
+      printf("Error checking for existing Eggdrop process.\n");
     }
+    fclose(f);
   }
 
   /* Move into background? */
@@ -1151,9 +1155,15 @@ int main(int arg_c, char **arg_v)
     setpgid(0, 0);
 #endif
     /* Tcl wants the stdin, stdout and stderr file handles kept open. */
-    freopen("/dev/null", "r", stdin);
-    freopen("/dev/null", "w", stdout);
-    freopen("/dev/null", "w", stderr);
+    if (freopen("/dev/null", "r", stdin) == NULL) {
+      putlog(LOG_MISC, "*", "Error renaming stdin file handle: %s", strerror(errno));
+    }
+    if (freopen("/dev/null", "w", stdout) == NULL) {
+      putlog(LOG_MISC, "*", "Error renaming stdout file handle: %s", strerror(errno));
+    }
+    if (freopen("/dev/null", "w", stderr) == NULL) {
+      putlog(LOG_MISC, "*", "Error renaming stderr file handle: %s", strerror(errno));
+    }
 #ifdef CYGWIN_HACKS
     FreeConsole();
 #endif
@@ -1162,6 +1172,9 @@ int main(int arg_c, char **arg_v)
   /* Terminal emulating dcc chat */
   if (!backgrd && term_z) {
     int n = new_dcc(&DCC_CHAT, sizeof(struct chat_info));
+
+    if (!n)
+      fatal("ERROR: Failed to initialize foreground chat.", 0);
 
     getvhost(&dcc[n].sockname, AF_INET);
     dcc[n].sock = STDOUT;
@@ -1179,7 +1192,8 @@ int main(int arg_c, char **arg_v)
       dcc[n].user = get_user_by_handle(userlist, dcc[n].nick);
     }
     setsock(STDOUT, 0);          /* Entry in net table */
-    dprintf(n, "\n### ENTERING DCC CHAT SIMULATION ###\n\n");
+    dprintf(n, "\n### ENTERING DCC CHAT SIMULATION ###\n");
+    dprintf(n, "You can use the .su command to log into your Eggdrop account.\n\n");
     dcc_chatter(n);
   }
